Clean up the wall class
