+++ lib/constants/parallax_constants.dart
@@ -1,4 +1,4 @@
 class ParallaxConstants {
   // How much we shift horizontally to the right as a fraction of moving down
-  static const horizontalDisplacementFactor = 1 / 12;
+  static const horizontalDisplacementFactor = 1 / 14;
 }
