+++ lib/constants/versioning_constants.dart
@@ -8,7 +8,7 @@ class VersioningConstants {
   static final String version = _getVersion();
 
   static const _majorVersion = 0;
-  static const _minorVersion = 9;
+  static const _minorVersion = 10;
   static const _patchVersion = 0;
 
   static const _releaseVersion = ReleaseVersion.alpha;
